fix code review

+++ src/main/java/com/froud/fraudservice/controller/FraudUserApiController.java
@@ -27,28 +27,28 @@ public class FraudUserApiController implements FraudUserApi {
     @Override
     public ResponseEntity<List<FraudUser>> getAllFraudUsers() {
         final List<FraudUser> fraudUsersList = fraudUserApiService.getAll().stream()
-                .map(fraudUserMapper::mapToFraudUser)
-                .toList();
+            .map(fraudUserMapper::mapToFraudUser)
+            .toList();
         return ResponseEntity.ok(fraudUsersList);
     }
 
     @Override
-    public ResponseEntity<FraudUser> getFraudUserById(final Integer id) { // почему то id -> Integer
-        final FraudUsersEntity fraudUsersEntity = fraudUserApiService.findById(Long.valueOf(id));
+    public ResponseEntity<FraudUser> getFraudUserById(final Long id) {
+        final FraudUsersEntity fraudUsersEntity = fraudUserApiService.findById(id);
         final FraudUser fraudUser = fraudUserMapper.mapToFraudUser(fraudUsersEntity);
         return ResponseEntity.ok(fraudUser);
     }
 
     @Override
-    public ResponseEntity<FraudUser> updateFraudUserById(final Integer id, final FraudUser fraudUser) {
-        final FraudUsersEntity fraudUsersEntity = fraudUserApiService.updateUser(Long.valueOf(id), fraudUser);
+    public ResponseEntity<FraudUser> updateFraudUserById(final Long id, final FraudUser fraudUser) {
+        final FraudUsersEntity fraudUsersEntity = fraudUserApiService.updateUser(id, fraudUser);
         final FraudUser fraudUser1 = fraudUserMapper.mapToFraudUser(fraudUsersEntity);
         return ResponseEntity.ok(fraudUser1);
     }
 
     @Override
-    public ResponseEntity<Void> deleteFraudUserById(final Integer id) {
-        fraudUserApiService.deleteById(Long.valueOf(id));
+    public ResponseEntity<Void> deleteFraudUserById(final Long id) {
+        fraudUserApiService.deleteById(id);
         return ResponseEntity.ok(null);
     }
 }
